% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{list_merge}
\alias{list_merge}
\title{Merge two lists and overwrite latter entries with former entries
if names are the same.}
\usage{
list_merge(list1, list2)
}
\arguments{
\item{list1}{list}

\item{list2}{list}
}
\value{
the merged list.
}
\description{
For example, \code{list_merge(list(a = 1, b = 2), list(b = 3, c = 4))}
will be \code{list(a = 1, b = 3, c = 4)}.
}
\examples{
\dontrun{
stopifnot(identical(list_merge(list(a = 1, b = 2), list(b = 3, c = 4)),
                    list(a = 1, b = 3, c = 4)))
stopifnot(identical(list_merge(NULL, list(a = 1)), list(a = 1)))
}
}

