% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tundraContainer-train.R
\name{train}
\alias{train}
\title{Train a model encapsulated within a tundraContainer.}
\usage{
train(dataframe, train_args = list(), verbose = FALSE, munge = TRUE)
}
\arguments{
\item{dataframe}{data.frame. The dataset to train the model on. This
will be preprocessed with the \code{tundraContainer}'s
\code{munge_procedure} and then passed as the first argument to
the \code{tundraContainer}'s \code{train_function}.}

\item{train_args}{list. A list of arguments to make available
to the \code{tundraContainer}'s \code{train_function} through
use of the \code{input} keyword. See the examples.}

\item{verbose}{logical. Either \code{TRUE} or \code{FALSE}, by
default the latter. If \code{TRUE}, then output produced by
running the \code{munge_procedure} or the \code{train_function}
will not be silenced.}

\item{munge}{logical. Either \code{TRUE} or \code{FALSE}, by
default the former. If \code{FALSE}, the \code{munge_procedure}
provided to the container during initialization will be assumed
to have been trained, and the \code{dataframe} provided will not
be run through it.}
}
\value{
The value returned by the \code{tundraContainer}'s
   \code{train_function}. Since the \code{train_function} has side effects
   on the container as its primary purpose, this can usually be
   \code{invisible(NULL)}.
}
\description{
Train a model encapsulated within a tundraContainer.
}

