% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/tundra_container.r
\docType{class}
\name{tundraContainer-class}
\alias{tundraContainer-class}
\alias{tundra_container}
\title{Tundra container class}
\description{
TODO: Formally define parameter spaces for models
}

