% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tundraContainer.R, R/tundra_container.r
\docType{class}
\name{tundraContainer}
\alias{tundraContainer}
\title{A standard container format for classifiers developed in R.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tundraContainer
}
\description{
A standard container format for classifiers developed in R.

TODO: Formally define parameter spaces for models
}
\keyword{datasets}

