% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tundraContainer.R, R/tundra_container.r
\docType{class}
\name{tundraContainer}
\alias{tundraContainer}
\title{A standard container format for classifiers developed in R.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x10da0d3c0> 
 - attr(*, "name")= chr "tundraContainer_generator"
}}
\usage{
tundraContainer
}
\description{
A standard container format for classifiers developed in R.

TODO: Formally define parameter spaces for models
}
\keyword{datasets}

